Simplified the "IP" (input) class and made it non-static.

+++ Source/PictureBox.cpp
@@ -42,7 +42,7 @@ PictureBox::PictureBox(Scene* scene, const sf::Vector2f& position, const int ID,
         else biggerNum = textureSize.y;
         outlineOffset_ = sf::Vector2f{(biggerNum / 300), (biggerNum / 300)};
     }
-    else outlineOffset_ = sf::Vector2f{options->GetOutlineSize(), options->GetOutlineSize()}; //or just use the outline size that was set in the options
+    else outlineOffset_ = sf::Vector2f{(float)options->GetOutlineSize(), (float)options->GetOutlineSize()}; //or just use the outline size that was set in the options
     outline_.setSize(sf::Vector2f(imageWidth_ - 2 * outlineOffset_.x, imageHeight_ - 2 * outlineOffset_.y));
     outline_.setFillColor(sf::Color::Transparent);
     outline_.setOutlineThickness(outlineOffset_.x);
